% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlpProtocol.R
\name{createPlpProtocol}
\alias{createPlpProtocol}
\title{Create a protocol template for the study}
\usage{
createPlpProtocol(predictionAnalysisListFile = NULL, outputLocation = getwd())
}
\arguments{
\item{predictionAnalysisListFile}{the location of the json analysis settings}

\item{outputLocation}{Directory location where you want the protocol written to}
}
\description{
Create a protocol template for the study
}
\details{
This function will create a template protocol
}
