% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeletonPredictionStudy.R
\docType{package}
\name{SkeletonPredictionStudy}
\alias{SkeletonPredictionStudy}
\title{SkeletonPredictionStudy}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
